app/requirements.txt
fastapi==0.88.0
python-dotenv==0.19.0
urllib3==1.26.9
uvicorn==0.20.0
websockets==10.4
pydantic==1.10.2
hermetrics==0.1.13
python-jose[cryptography]
passlib[bcrypt] 
python-multipart
ipython
